add sort_currencies
fix router latent ordering

encode_v4_exact_in_single_params currency params retyped str -> Address to match what they already receive at runtime
